[new-env] Record reads precisely and avoid unnecessary reads in member expressions
Summary:
In D35519898 (https://github.com/facebook/flow/commit/9f68b8210ddff61bb28dbaa262a534ce7944144b), I add an expression read before `member_expression_refinement`, so that the previously refined value can be read. This seems to be a safe operation to do, but it has serious side effects. For example, if the member expression is `foo.bar.methodCall().baz` and `foo.bar` is a refined value, doing so will invalidate the refinement, and a later read of `foo.bar` in `member_expression_refinement` will read the unrefined value of `foo.bar`.

Therefore, we need to be more careful to avoid the side effects of `this#expression`. In this diff, I replaced that with `record_member_read` factored out from optional chain. `record_member_read` will only try to get the val and record the read, so it won't accidentally invalidate refinements.

There is a similar problem with `this#expression` call in `maybe_sentinel`. However, the call can be safely removed in that case, because every call of `maybe_sentinel` has already read the expression somewhere:

- `null_test`, `void_test`, `literal_test` in call of `this#optional_chain`
- `member/optional eq sentinel_value` in call of `this#expression`

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35706511

fbshipit-source-id: a56db2a2fb9f25a55a3f33a3bd278f5bdcbf8fff