[new-env] Connect literal test and optional chaining test into a single connected conjunction
Summary:
We need to put literal test, sentinel refinement, and potential non-maybe refinements in optional chain into a single connected conjunction, so that when sense is false, we will also negate the refinements in optional chain.

e.g. in `obj?.prop !== 'foo'`, the refinement will now correctly be `Or(Maybe(obj), Not Sentinel prop, Not Literal foo)`.

It might not seem very useful, but it's more powerful when it's negated. For example, a lot of graphql code might do:

```
if (!node?.__typename !== 'some type name') throw some error;
// now node is definitely not null!
```

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35668475

fbshipit-source-id: 1131d2e131a63ee35574057f47a82870692d22a4