remove Sys_utils functions that crash
Summary:
there are a bunch of `*_no_fail` functions in `Sys_utils` that catch all exceptions, log to stderr and then `exit 3`. this doesn't go through our standard error loggers or use `FlowExitStatus` (code 3 means "Out of time" according to that!).

instead, just let them throw exceptions. we have a catchall handler in `flow.ml` that will catch this and log it and exit with `Unknown_error`.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D35778067

fbshipit-source-id: 97c6253d97d2be21941580177204f7c12669c973