[new-env] Use `def_loc` to get type of object in matching props to match the old env behavior of using general type of object
Summary:
Old env only supports matching props of the form `name.prop === literal`. We once supported more complex form of object expressions in the new env, but it's causing too much differences so we disabled it in D35586894 (https://github.com/facebook/flow/commit/cdd63c8b2bbe49c9944704d51ba6da0eef0cab5d).

With that change in place, it's now possible to get rid of the final difference: the old env will use the general type of the `name`, while the new env uses the type of `name` at the location of comparison. In the new env, we get the general type by using the def_loc of `name` and then get the provider type of the def_loc during post inference check, in the same way we do for literal subtype checking.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D35822022

fbshipit-source-id: 4340a77eacd490910cc78523a18c6784d0427432