Fix haste module resolution
Summary:
Before this change, when resolving a haste module, we would first check for Node
modules.

Checking for Node modules is relatively quite expensive, since it requires stat,
readdir, and realpath calls, which are particularly slow on EdenFS.

Beyond perf, this behavior is also just wrong. If we have a Haste module called
"foo" and a node module "foo", require("foo") should resolve to the Haste
module, not the Node module.

Following Metro, the order is:
1. haste module
2. haste package (magic dir)
3. node modules

Reviewed By: mroch

Differential Revision: D35827588

fbshipit-source-id: 3f05884a2fd6eb19aa2fac076632a00d8f1bf7fc