remove file_exists check from Module_js.Node.parse_main
Summary: I don't think we need this check. when it was written, we would error if we tried to look up a package that wasn't in the heap. these days, `get_package` returns `None` and all is fine. we should avoid hitting the disk here; it's cheaper to test the shared hash table than to do a `file_exists` check on eden.

Reviewed By: samwgoldman

Differential Revision: D35805594

fbshipit-source-id: 7b9480546bb83a21729f0c4e42f477025146e6d0