Autocomplete members of recursive types
Summary: Currently if one tries to autocomplete a jsx property on a component whose type is recursive, we'll produce nothing. However, it seems reasonably safe to allow `Mu` types to be explored for the purposes of autocompletion: while we may end up with type variables (`V$1`) in the type information display, which is unfortunate, it's probably a better tradeoff to allow autocompletion to happen at all.

Reviewed By: mroch

Differential Revision: D35860403

fbshipit-source-id: 9047819d4da6a223e7af48d70d09e8cde7643d1c