[constrain_write] Move constrain write check to the top
Summary:
## Analysis

The post inference checks are performed when we have already inferred all the types. We would want to have the property that all these checks are independent and never establishes new constraints. It seems reasonable to assume that's true, since we resolve the type before having more flows on them.

Unfortunately the claim is not true. In some post inference checks like matching props check, we use `resolver_visitor` class which only solves an OpenT shallowly. There is a TODO that says we should switch to `Context_optimizer`, which recursively resolve all tvars. This is definitely a better fix and I am working on it, but the new env is already relying on the behavior that some OpenT are left unresolved, which needs to be fixed first.

Meanwhile, we can move constrain write (which uses `Context_optimizer` check up, so that it won't be affected by other checks that uses `resolver_visitor`.

## Example

In case you wonder, here is a concrete example when unresolved tvars is causing problems

```
let x = [];
x = [3]
if (x.length === 1) {}
```

This is one example when two post inference checks overlap: both constrain writes and matching props check will need the type of `x`. In matching props check, `x` has type `Array<OpenT> | Array<number>` . This triggers a speculation for union, which during the speculation, `OpenT` got resolved to `any`. Therefore, the constrain writes check will see the same `OpenT` as resolved, and won't resolve it to `empty` as we hoped. This causes the missing constrain writes error.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D35800961

fbshipit-source-id: a0974ff8c504d75d5ea3d4d049b8920cf6635d87