Fix try-catch-finally
Summary:
This diff fixes how the name_resolver handles try-catch-finally statements.

The behavior here also much more closely matches what we do in statement.ml. In particular, the part where we check the finally twice is the same approach taken in statement.ml.

So now, an overview of the changes:
We only change how we traverse the `catch` and `finally` blocks-- the `try` block is unaffected here.

## Catch
Before this diff, the entry state of the catch block was modeled as the merge of the try_entry and try_exit. This was close, but not quite correct, as it did not take into account the explicit throws from the `try` block. If there's an explicit throw in the try block then the `try_exit` would be empty, and we need to call `commit_abrupt_completions_matching` with Throw to merge in those writes. So we do that here.

## Finally

There's two interesting parts to finally:
1. What should the post-state of the try-catch-finally look like?
2. What should the env be inside the finally?

It turns out that it's hard to answer both of these questions with just one traversal. If the catch block throws, then the state after the finally is unreachable. If the try or catch block does not throw, then the state after the finally would be reachable unless the finally also abruptly completes. So we check twice. The first check we do assumes that either try or catch do not throw, with a special case for when we know the catch definitely throws. This tells us what the env should be after processing the finally. Then we check a second time without making any assumptions about how the try/catch behave to actually type the variables in the try block, since making an assumption about the behavior of the try or catch there would be unsound. See the inline comments, which explains this strategy as well.

See https://github.com/facebook/flow/blob/main/src/typing/statement.ml#L1388-L1473 for statement.ml's approach.

Reviewed By: SamChou19815

Differential Revision: D35825649

fbshipit-source-id: bb8347d64f0b04cc249f5bae8075a078ef74d0c1