Do not merge heap refinements after conditionals if they are over the same value
Summary:
Conditional branches over heap entries should not create a merge of the positive and negative case at the end. For example:

```
if (x.foo === null) {}
// Should not be a PHI of === null and !==null, since that is just the same as x.foo
x.foo;
```

In the best case, making that merge is just a complicated way of recomputing the original type. In the worst case, the heap refinement could be introducing a value that does not actually exist, and this behavior masks future GetPropT failures. For example:

```
  const a = {p: ''};
  (a.q ? a.q : '');
  a.q; // no missing prop in new-env
```

The object_widening test contained a case where this was hit, so I didn't add an additional test case

Reviewed By: SamChou19815

Differential Revision: D35845360

fbshipit-source-id: 51981b250e0fd0b83bf92743cd26d2a180699f9a