[new-env] Cache all Val.one
Summary:
Changelog: [internal]

`Val.one` represents a unique write location. Therefore, the following should be true

> ∀ v1 = Val.one loc1, v2 = Val.one loc2, loc1 = loc2 implies that v1 = v2.

This is a nice property to have, but it's currently not true. We currently create different `Val.one` for the same unique write in the following places:

1. During `mk_env` when we create the val for `havoc`, and then during `pattern_identifier` when we replace potentially undeclared `val_ref` with `Val.one`.
2. During joins in `Val.merge` and `Val.all`.

This is not a problem so far, but it starts to become a problem when we want to compare a val with its corresponding havoc. Currently, they will *never* have the same SSA id even when they should be conceptually the same. e.g.

```
const a = 1; // havoc val: ssa_id=1, val_ref: ssa_id=2
a; // read ssa_id=2, compare with havoc and found they different id even though they should be the same.
```

This diff solves the problem by caching the reason for Val.one to ensure what we only create one unique Val.one for every unique write loc.

Reviewed By: panagosg7

Differential Revision: D35850144

fbshipit-source-id: 4a2d9b2becfd2153b05cf8ba1d476df721d817d4