Remove merge timeout code
Summary:
Since types-first 2.0, the merge phase does very little and is not a performance
concern. This code was originally added before types first, when merge included
all type checking. Checking now happens after merge, in the check phase, which
has its own timeout code.

Removing this code is a nice simplification to the code, building on the
previous diff in this stack, continued in the next.

Reviewed By: mroch

Differential Revision: D35860236

fbshipit-source-id: 0bfad33bfe2c99e19a8c96c0d2169fce463ecbe7