[contextual] do not propagate type hints to arguments of special cased functions
Summary:
Before this diff in "enforce_local_inference_annotations" mode we would not require
an annotation for `x` in
```
Object.assign({}, { foo(x) {} });
```
assuming that we would be able to contextually produce such an annotation based
on the receiving method type. For library functions like this (which are special cased)
in the checker, we cannot expect to provide a reasonable type for the parameters of
`foo` and so an annotation for `x` should be required.

I'm making this change in a separate diff in preparation of D33338700 that takes a more
thorough approach to how hints are propagated, so that the latter has a neutral error
overhead.

Reviewed By: SamChou19815

Differential Revision: D35834604

fbshipit-source-id: 77aeb5a50fadac3f1de81a35c4baf9c3fc56b32e