Classes in resolved env
Summary:
Support for real classes in the resolved environment. There are two cases here: fully annotated classes, where we can extract a signature without checking any method bodies, and ones where inference is needed. The name_def analysis will have already distinguished between these cases, so we now just need to consume them separately.

In keeping with most of this line of work, this is relatively straightforward: we expose the `mk_class` (for unannotated classes) and `mk_class_sig` (for annotated classes) functions from statement.ml and invoke them directly on the ASTs for the classes, which we've previously gathered.

As of this diff we don't cache checked classes, which means that when we typecheck the overall program we will end up re-checking (and generating different tvars for) class implementations. The remainder of this stack fixes that, but requires major refactoring so it happens over the course of several diffs.

Reviewed By: panagosg7

Differential Revision: D33501740

fbshipit-source-id: 5c05365752869f1b3bc674d8c7d3897f7b73938a