[contextual] add hint type
Summary:
In contextual typing we often need to perform decompositions on type hints before
we descend deeper into an expression during type checking. For example if an
object literal
```
{
  f: (x) => x + 1
}
```
is checked with a hint
```
{ f: (number) => number }
```
then we can check the value of property `f` with the hint `(number) => number` and,
further, use `number` as the type of `x`.

There is a few kinds of decompositions on hints we might be interested in. In the above
example the first operation is selecting a property from an object type, and the second
is extracting a parameter type from a function type.

This diff introduces all kinds of decompositions we are interested in while checking a
program, and enhances the type of `~hint` used in statement.ml from being just a `Type.t`
to also include decompositions over types. We could compute these decompositions
as we recurse into expressions, but that might lead to unnecessary computations since
the hints might not be needed after all.

In a follow up diff we use these hints in the recursive traversal of statement.ml.

Reviewed By: SamChou19815

Differential Revision: D35801943

fbshipit-source-id: b6861e33c97cb20f1526d2f7f8d76f9233c87f79