[contextual] decompose scaffolding
Summary:
This is a mostly mechanical diff. It propagates a `~hint` label in various recursive
functions in statement.ml. This label preexisted for a few methods and had type
`Type.t option`. Note, however, that the type of the hint was not correctly
decomposed as we descended to the various kinds of expressions. In fact this was
a TODO in the form of calls to `hint_decompose_opt_todo`.

This diff changes the type of `~hint` to be `Type_hint.hint` -- a type introduced in
D35801943. This type encodes the necessary information to produce the appropriate
hint based on what expression we're recursing into. Note that the actual decomposition
operations are still future work. So the manual portion of this diff is to select the right
variant of `Type_hint.hint_decomposition` for the various calls of `expression` in
statement.ml.

This diff is not expected to have a visible change in behavior, and is expected to go
through various iterations, as hint computation becomes more precise.

Reviewed By: SamChou19815

Differential Revision: D33338700

fbshipit-source-id: ecb52d51367469761434fca9c3073beb5dfff7f9