`instanceof Α` refines any to an instance of Α
Summary:
Following D14709211 (https://github.com/facebook/flow/commit/159c4746e402df34176417261d9093f81d93704d) that handles the case of `mixed`, `x instanceof A`,
when `x` is `any`, refines it to `A` (or `A<any, ...>` if `A` is polymorphic). Before,
the type of `x` after the refinement was `empty`.

Changelog: [fix] `x instanceof A` when `x` is typed as `any` refines it to `A` instead of `empty`

Reviewed By: gkz

Differential Revision: D20808352

fbshipit-source-id: 48d5ef8a560c086d205501d06111d1c45b33978c