fix crash when resolving requires
Summary:
suppose file A depends on B.

- change B. calc_direct_dependents caches B -> A
- introduce a parse error in A. the recheck changes A to untyped
- change B again. we hit the cache for B and think that A is still a direct dependent

A is NOT a direct dependent anymore, because it doesn't parse (or is deleted).

we pass A into `resolve_requires`. `resolve_requires` assumes it's only passed typed files and throws a `Parsing_heaps.File_not_typed` exception, killing the worker, and thus killing the server.

the fix is to clear the cache when a file becomes untyped or deleted.

Changelog: [fix] Fix a crash when deleting a file or saving a file with parse errors in certain circumstances

Reviewed By: samwgoldman

Differential Revision: D35955111

fbshipit-source-id: 92a0e0aca1b409b48574377932c1efd1c012fcb1