fix missing parse errors in objects
Summary:
fixes a regression in 0.175.0 where the parser misses errors when an object property has a value but the semicolon is missing.

D34885414 (https://github.com/facebook/flow/commit/aca5c1114f08521a5d641a9a5a514421d2f54708) improves error recovery when parsing incomplete properties in object literals and object patterns. to reduce duplicate errors, it didn't error when the semicolon was missing since it said so in a different error. but only when there's no value.

the custom error is nice but it's tricky: there are lots of other things it could be expecting -- (, < or =, and it depends on whether it's a pattern or an expression. too much code for this one case, so i went back to the basic "expected ,". it'd be nice to handle this generically if the parser could know all the tokens it's trying to peek at.

Changelog: [parser] Fix a parser bug that allowed missing commas between object properties (regression in 0.175.0)

Reviewed By: gkz

Differential Revision: D35881308

fbshipit-source-id: 93e64030875ee8dbc06d19fd815ce7158f0b6aa0