Tease apart type definitions for class and function sigs
Summary:
In D33501740 (https://github.com/facebook/flow/commit/fcca5bcb0138a39ef89d013758b006477cbcdde9) we introduced classes into the fully resolved environment, but we didn't try to cache them yet. The reason for this is that the class sig data structure (which is the infomation we would cache for fully annotated classes) are currently defined in `class_sig.ml`, which is in a fairly complex system of functor modules that are in recursion with statement.ml. The functor modules are needed because class sigs depend on function sigs, and we want to reuse our code for function sigs to handle both annotated and inferred functions--hence different definitions for function parameter signatures for type_annotation.ml and statement.ml. The upshot of this though is that if the node cache directly contains class sigs in their current form, it would have to be part of the recursion with statement.ml--but we want to *store* the node cache in the context, which is not part of this recursion.

This is the first of several diffs that fix this by significant refactoring. The observation here is that while the functions provided in class_sig.ml, which are used for creating and manipulating class sigs, depend on statement.ml, the actual type of signatures themselves don't. We can therefore separate this module into two modules: one that contains type definitions and one that contains functions.

However, since the types of class sigs (as well as the functions that manipulate them) are abstracted over different types for function sigs, in order to make this change, we need to do the same thing to function sigs, and in turn to function param sigs. This diff executes this refactoring, splitting each of the sig functors into ones for types and ones for implementations. The implementation functors depend on both the types modules and the lower-level implementations modules (e.g. class sig functions depend on class sig types and function and function param sig functions), while the types modules only depend on the lower-level types modules.

We don't do the rest of the work of separating these out into different TARGETS yet, or changing what modules they exist in, for the purpose of making this easier to review: that happens in the next diffs.

Reviewed By: panagosg7

Differential Revision: D35830774

fbshipit-source-id: c51d86cf4219a97a1d3d78890ace742793d9bc13