Fix finally entry env
Summary:
This diff fixes the entry-state for the `finally` block. Previously we used the merge of the try start env and the end of the catch env, which assumes a throw in try. Instead, we should merge the try start env, try end env, and catch end env + all the abrupt completions from try and catch.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D35975437

fbshipit-source-id: 86c762b5cce9cbd9045c82ea4e8aeeea95fa78ac