fix Path_matcher on Windows
Summary:
`Path_matcher.fixup_path` takes a path that may contain globs, and tries to normalize `.` and `..` out of the middle of it.

for example, if your `[includes]` contains `../foo/**/bar` then the callers of `Path_matcher.add` will pass in `/path/to/root/../foo/**/bar` and we want `/path/to/foo/**/bar`.

this was broken on Windows in 2 ways:

1) when the path was absolute, it prepended `/`, even on Windows. further, `Filename.concat` thinks that `:` is a dir separator for whatever reason, so `Filename.concat "C:" "foo"` results in `"C:foo`, not `C:\foo`. so `C:\foo\..\bar` resulted in `\C:bar`... 🤦‍♂️

2) Windows can use either `/` or `\`. the includes are always `/`, but we were only splitting on `\`. so after fixing (1), `C:\foo\../bar` remained as `C:\foo\../bar` instead of `C:\bar`.

Changelog: [fix] Fix an issue with `[include]` entries containing both globs and relative paths on Windows.

Reviewed By: samwgoldman

Differential Revision: D36007101

fbshipit-source-id: fef4f0ec5865d35e4d74a2b6c4fe2df7f23fb5fe