[new-env] Error on rebinding a declared function with var, let, const bindings
Summary:
Currently, we produce no error when declared function is rebind. A declared function can be rebind with another declared function with the same name, and rebind once with a concrete function declaration. It cannot be rebind with var, let, and const bindings.

Previously, we are unable to distinguish between a function binding and a var/let/const binding. The previous diff finally fixes the problem and give us the power to error on the case when a declared function is rebind with var, let, and const.

Note that we still don't error when the declared function is rebind with two concrete functions, e.g.

```
declare function foo(): void;
function foo(): void {}
function foo(): void {} // should error, but we currently don't
```

But it's a pre-existing problem in the old-env, and we can fix it later and don't let this block new-env rollout. I marked this as a TODO in the test case.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D36016072

fbshipit-source-id: 348b4bfa9e60335742ec8f17257791f6b84c8045