[new-env] Introduce `pattern_write_kind` in `name_resolver`
Summary:
Currently, we rely on the `Flow_ast.Statement.VariableDeclaration.kind` to tell us what's the type of binding of an identifier from AST. We also overload the kind by treating `None` as an assignment. This can be potentially confusing for someone who's reading the code for the first time. It also has the problem that we can't tell whether an identifier is a function identifier, since we hardcode it to be `Flow_ast.Statement.VariableDeclaration.Let`.

This diff makes everything more explicit by creating the `pattern_write_kind` variant that includes

```
    | VarBinding
    | LetBinding
    | ConstBinding
    | FunctionBinding
    | AssignmentWrite
```

Now we finally have a way to differentiate between a function identifier binding and a let binding.

In the next stacked diff, I will use this to error when a declared function is rebind with a var/let/const binding.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D36016073

fbshipit-source-id: 8ed69f6ca77044a401060e56e147ab5099ed2ec0