[new-env] Reposition the type of RHS of instanceof
Summary:
In the old env, the RHS of instanceof will be type checked and the type will be passed into the predicate. The type of RHS will always have the same loc as the loc of RHS.

This diff does a repositioning, so that the locations of type reasons are consistent with the old env.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35989576

fbshipit-source-id: 3ebb6cd89a9b3e881df24741784d93083b8fbb2e