[new-env] Model exports as a special val instead of global in new-env
Summary:
In the old env, the type of exports variable is an OpenT with the exported values dict flow to it. The variable is stored in a scope entry as an ordinary name, which is not a code path the new env will go through. As a result, the type of exported values will flow to the global any type, which is useless.

This diff models exports as a special module-scoped variable. Unlike other module scoped variables, it's not any-typed, but will get its type from `type_inference_js.ml`.

This change not only makes the `exports` have an accurate type for type_at_pos, but also enforces that the exports dictionary is contravariant. D33167270 (https://github.com/facebook/flow/commit/b538aca6174b2ee506d0c9323f88b9617a22c7ed) which first adds exports to `prelude.js` can now be reverted.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35975693

fbshipit-source-id: aa3fb22a5c194a5f27ce368c357267a93edf85b9