[new-env] Bool literal subtyping test should check for subtyping instead of always false comparisons
Summary:
Although the literal subtyping test also serves the purpose of detecting those always false conditions when the subtyping relation isn't true, it's error messages is designed for subtyping only. Thus, in the case of `foo !== false` when type of foo is `false`, we will get some confusing error like `true` is not compatible with false.

This is very confusing. Let's just restore the behavior like in the old env, and maybe consider add this kind of always true/false condition check later with more specialized error messages.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D36050030

fbshipit-source-id: 2d1aff7348fe0703f79f0125496a71085260f87d