[new-env] Fix unsoundness of logical refinement merge
Summary:
Currently, the way we merge OR and AND in refinements can lead to unsoundness issues.

Consider the following examples:

```
a != null && expr_introduces_no_refi
a != null || expr_introduces_no_refi
```

In the current implementation, they will both be normalized to `a != null`. For the case of `&&`, this is fine for the positive case, but when it's negated and normalized it becomes `a == null`, but instead it should really be normalized to nothing. For the case of `||`, it's unsound because we should have no applied refinement, but instead we get `a != null`.

This diff will merge them to `And(total, Refinements empty)`, `Or(total, Refinements empty)` instead, which preserves the original structure that one side has no refinements

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D36057541

fbshipit-source-id: 378a2cd0373e765b7195d55f3df97323c6798790