[new-env] Record member read instead of full expression visit in `binding_pattern_track_object_destructuring`
Summary:
Without the change, in the case of

```
const {a, b} = obj.prop();
```

We will call

```
this#expression 'obj.prop()'
this#expression 'obj.prop().a'
this#expression 'obj.prop().b'
```

During the second call, refinement on obj.prop will be invalidated, and we will get spurious errors.

What we need is just to recursively record the projection read on the synthetic expression.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D36050031

fbshipit-source-id: 4cacd41b0a04a6d62ba476266cb9b084b0173925