[new-env] Use havoc value as merge result for certain val merges
Summary:
In the old env, we have the shortcut of type merging when the specific type in one branch is equal to the general type:

https://github.com/facebook/flow/blob/6e59c775ac0f4ab3462fb7cd8a34d5b139e95271/src/typing/env.ml#L1384-L1386

Currently, the new env does not have this shortcut and is more precise as a result. However, the precision is not always desirable. Sometimes, it even has the effect of ignoring user's type annotations:

```
const foo: Array<string> = [];
let bar: Array<any> = foo;
if (true) bar = foo;
(bar: Array<boolean>); // ok in old env, error in new env, because we merge too precisely
```

This diff matches the old env behavior with best effort. It might seem like it's a direct translation:

```lang=ocaml
(* from *)
if specific1 = general0 || specific2 = general0 then general0
(* to *)
if havoc = v1 || havoc = v2 then havoc else Val.merge v1 v2
```

but it's not that simple due to the fundamental difference that old-env merge operate on types and the name resolver merge operates on syntactic values. Therefore, the following additional restrictions have to be put on the name resolver merge:

1. The havoc value must be annotated. The old env is heavily tvar-based and the general type can be widened as a result of type checking individual branches. Without the restriction, we might incorrectly miss some widening of the type as desired. There is no way to know that in name resolver, so we take the most conservative approach to only use havoc value when it's annotated, so we know it cannot be widened.
2. The other value that's not equal to havoc cannot be uninitialized. The old env handles the uninitialized variable problem by keeping track of the initialization state in scope entry, while in the name resolver they are separate vals. Therefore, we must fall back to use `Val.merge` to not miss the uninitialized variable case.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D35689193

fbshipit-source-id: 150ae6228ad4f405e1bf31c00dd9bba3cafc0f61