Support patterns in builtin type sigs
Summary:
I assumed that Flow did not support destructuring patterns in libdefs, which was
completely wrong. This led to crashes when extracting signatures for libdefs
containing patterns.

The fix is simple: use the existing support for patterns in source files for
builtins.

Fixes https://github.com/facebook/flow/issues/8874

Changelog: Fix crash when declaring a destructured binding in a library definition

Reviewed By: gkz

Differential Revision: D36073585

fbshipit-source-id: 368b0171e6d5d1b574839269a9a3495f8af4b5d1