Add of_increasing_iterator_unchecked
Summary:
When we have a sorted collection and want to create a balanced tree, this API
avoids all comparisons and rebalancing operations that we might normally do. As
the name suggests, this function assumes that the input is actually sorted.

I plan to use this to optimize loading saved states, later in this stack.

Reviewed By: mroch

Differential Revision: D36041631

fbshipit-source-id: ce9bf39f0bde0e8ecb665f029d95d70f3d530f74