Stop using File_key.Builtins during libdef init
Summary:
Library definitions are inferred with a shared component context, which is then
optimized and stored in shared memory. (This is very similar to how merge used to
work before types-first.)

Before this diff, we would create a "master_cx" Context.t, sharing the component
context, and used this as the handle to optimize and store the builtins. This
context used the File_key.Builtins variant.

After this diff, we don't create this handle at all. Instead, we just use one of
the lib file contexts, which use the File_key.LibFile variant. There's an edge
case of no successfully parsed lib files, and therefore no lib file contexts, in
which case we just create an empty master context.

This is the second and last use of File_key.Builtins. The next diff will remove
the variant definition.

Changelog: [internal]

Reviewed By: mroch

Differential Revision: D36103697

fbshipit-source-id: 2cf7ea813c65997a78bf1aa2bd0deafc95a45e73