[new-env] Do not return a bare projection in `get_val_of_expression`, and union changeset
Summary:
Changelog: [internal]

## Motivation

Changesets represent newly created projections for heap refinements in scoped refinements. We only add (lookup, projection val) to the changeset when the projection val is *first* created. This works well in many cases, and allow us to kill the projection when the changeset out of scope, and allow us to not lose any outer scope refinements that has the same syntax of lookup.

e.g.

```
if (x.a) {
  if (x.a) {
  }
  x.a // refinement from the outer if should be still alive.
}
```

However, this messed up the case like `obj.prop != null || obj.prop != null` (While this example looks silly, it's just a simpler repro of more legitimate code like `(obj.prop != null && flag === 'Foo') || (obj.prop != null && flag === 'Bar')`.) In this case, the changeset is only created on LHS, then when we compute changeset in `normalize_total_refinements`, we take the intersection for `Or`, which causes the changeset to be dropped completely.

On another seeming unrelated example, consider

```
if (data.foo.bar?.p == null && data.foo.bar?.q == null) throw '';
// data.foo.bar?.p == null produces refinement:
//   !(data.foo.bar != null && data.foo.bar.p != null)
```

However, due to negation, the above will be turned into a disjunction with two changesets of different lookup keys, which normalized to nothing. This causes us to unable to reuse the same base projection `data.foo.bar` on the RHS, which leads them to have different SSA IDs. Then when we are in the else case and negate the refinement again, we have:

```
(data.foo.bar != null && data.foo.bar.p != null)
|| (data.foo.bar != null && data.foo.bar.q != null)
```

Since two `data.foo.bar != null` have different SSA IDs, both will be lost in an OR merge!

## Changeset's conflicting responsibility

All of the above seem to suggest that we should union changeset. However, changeset also has the responsibility to remove projection in heap refinements when the refinement falls out of scope. For that purpose, we must intersect changesets when merging ORs.

Therefore, the changeset is responsible for both removing and restoring base projections. Removing projection requires intersection, and restoring projections requires union.

## Solution

However, it's important to step back and realize that these two conflicting responsibilities do not need to be fulfilled by changeset.

In particular, the unrefinement process in `pop_refinement_scope` can already completely kill refinements that's out of scope. The only missing piece is that we still have a bare projection in heap refinements that should be removed. This can be achieved by never returning a bare projection in `get_val_of_expression`. After that, we can union changeset without any safety concern!

There is only one catch with this approach. We currently add a "permanent" projection for logical assignments. We do that by doing it outside of changeset mechanism, so it won't be killed during `pop_refinement_scope`. Currently, we add projections to it, but due to the change in this diff, the projection will be ignored when we query in `get_val_of_expression`, so I change it to store a `Val.one` instead.

Reviewed By: jbrown215

Differential Revision: D35837249

fbshipit-source-id: a39cddf7f2316d23b345186de1b5e2705f7dbece