Do not throw from new-env in prod mode
Summary:
We should only crash on a missing entry for the new-env in build mode-- in prod mode we should instead emit an error and return any so that we do not crash on a single unhandled pattern.

This diff wraps the function calls in new_env.ml that may cause an exception in a wrapper that catches the exception and returns `any` with an error instead of crashing in prod build mode. This will make the new-env much more robust when it's used out in the wild.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D36077306

fbshipit-source-id: 39f23bf332d9098efc34662380450c6521d488df