fix file finder with overlapping stems
Summary:
`Path_matcher` encapsulates all of the `[include]` regexes. it extracts "stems": actual paths at the beginning of the regexes, before any wildcards. an include path like `/foo/bar/**/*.js` has stem `/foo/bar`.

`Files.make_next_files` walks all of the stems, checking whether each file is included (passes `Path_matcher.match`).

### bug

when there are overlapping stems, it walks files multiple times!

for example, see the `libs_are_implicitly_included` test. it has `../lib.js` in `[libs]`; since the root is implicitly included, we end up with `/path/to/root`  and `/path/to` as stems.

`make_next_files` walks all of the stems: it walks all of the files in `/path/to/root` and sees that yes, `/path/to/root/a.js` matches *some* regex (`/path/to/root/**`).

then it walks all of the files in `/path/to` and sees that yes, `/path/to/root/a.js` matches some regex (again, `/path/to/root/**`). so we visit that file twice!

### fix

sort and dedup overlapping stems

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D36135674

fbshipit-source-id: a29886ecd4c8329934699ec0eeb4f1f66cfa726c