Add flag for contextual typing
Summary:
Before we are able to wire together contextual typing, fully resolved environment and implicit instantiation, it will be expensive to perform hint decomposition for contextual typing, because we need to do a runtime check to ensure that the base hint is fully resolved.

Therefore, we need to gate contextual typing for now. This diff adds the flag.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D36137462

fbshipit-source-id: 3f68a285acae4aed7e5f929b2d47bb30bda525f7