Havoc loop changed vars on different ssa ids
Summary:
The current check for if a value is changed in a loop fails to catch cases like the one in the added test case.

This diffs falls back to a more simple and conservative approach, which considers a variable to be changed if it's ssa id is different.

This will make cases like this havoc:
```
let x: number | string = '';
invariant(typeof x === 'string');
while (true) {
  if (typeof x !== 'string') { break; }
  x; // x's ssa id is changed here, so we'll havoc before entering.
}
```

but it seems like this sort of issue happens very rarely in practice. We can revisit a more complete solution in the future.

In addition, this diff fixes several instances in which heap refinements were not correctly cleared at the end of a loop. See the changes to `post_loop_refinements` and updated comments.

Reviewed By: SamChou19815

Differential Revision: D36068841

fbshipit-source-id: ad90c78feb0f5492f4e397d49f3e62711f87bbdb