Phantom dependencies are modules
Summary:
Flow's dependency graph has an indirection through modules. Files depend on
modules, modules have providers.

A haste module's providers come from files' haste names. A file module is
provided by either a .js file or a .js.flow file.

A phantom dependency corresponds to a file which does not exist, but if it did
exist, would become a provider for a dependency module.

Before this change, we represented phantom dependencies as file paths. These
file paths, however, correspond to file *module* names. That is, we don't store
.js.flow files as phantom dependencies.

This diff recognizes that it is equivalent to consider phantom dependencies as
modules themselves. They are phantom modules, which have no provider file, but
would become a dependency if they did.

This allows us to simplify a few things:

1. When resolving requires, we can replace the case-sensitive lookup code with a
check to the shared hash table, which is case sensitive already.

2. When calculating direct dependents, we only need to consider changed modules
and not files.

It also enables the stacked fix, where we need to track a non-file phantom
dependency.

Reviewed By: mroch

Differential Revision: D36136081

fbshipit-source-id: dba9d7af1da38232163d45caf6cb030976f882e5